+++ examples/api/CMakeLists.txt
@@ -1,22 +1,13 @@
 set(CVC4_EXAMPLES_API
   bitvectors
-  bitvectors-new
   bitvectors_and_arrays
-  bitvectors_and_arrays-new
   combination
-  combination-new
   datatypes
-  datatypes-new
   extract
-  extract-new
   helloworld
-  helloworld-new
   linear_arith
-  linear_arith-new
   sets
-  sets-new
   strings
-  strings-new
   sygus-fun
   sygus-grammar
   sygus-inv
