+++ examples/api/bitvectors.cpp
@@ -1,8 +1,8 @@
 /*********************                                                        */
-/*! \file bitvectors.cpp
+/*! \file bitvectors-new.cpp
  ** \verbatim
  ** Top contributors (to current version):
- **   Liana Hadarean, Aina Niemetz, Morgan Deters
+ **   Aina Niemetz, Makai Mann
  ** This file is part of the CVC4 project.
  ** Copyright (c) 2009-2020 by the authors listed in the file AUTHORS
  ** in the top-level source directory) and their institutional affiliations.
@@ -16,15 +16,15 @@
 
 #include <iostream>
 
-#include <cvc4/cvc4.h>
+#include <cvc4/api/cvc4cpp.h>
 
 using namespace std;
-using namespace CVC4;
+using namespace CVC4::api;
 
-int main() {
-  ExprManager em;
-  SmtEngine smt(&em);
-  smt.setLogic("QF_BV"); // Set the logic
+int main()
+{
+  Solver slv;
+  slv.setLogic("QF_BV");  // Set the logic
 
   // The following example has been adapted from the book A Hacker's Delight by
   // Henry S. Warren.
@@ -46,81 +46,82 @@ int main() {
   // equivalent by encoding the problem in the bit-vector theory.
 
   // Creating a bit-vector type of width 32
-  Type bitvector32 = em.mkBitVectorType(32);
+  Sort bitvector32 = slv.mkBitVectorSort(32);
 
   // Variables
-  Expr x = em.mkVar("x", bitvector32);
-  Expr a = em.mkVar("a", bitvector32);
-  Expr b = em.mkVar("b", bitvector32);
+  Term x = slv.mkConst(bitvector32, "x");
+  Term a = slv.mkConst(bitvector32, "a");
+  Term b = slv.mkConst(bitvector32, "b");
 
   // First encode the assumption that x must be equal to a or b
-  Expr x_eq_a = em.mkExpr(kind::EQUAL, x, a);
-  Expr x_eq_b = em.mkExpr(kind::EQUAL, x, b);
-  Expr assumption = em.mkExpr(kind::OR, x_eq_a, x_eq_b);
+  Term x_eq_a = slv.mkTerm(EQUAL, x, a);
+  Term x_eq_b = slv.mkTerm(EQUAL, x, b);
+  Term assumption = slv.mkTerm(OR, x_eq_a, x_eq_b);
 
   // Assert the assumption
-  smt.assertFormula(assumption);
+  slv.assertFormula(assumption);
 
   // Introduce a new variable for the new value of x after assignment.
-  Expr new_x = em.mkVar("new_x", bitvector32); // x after executing code (0)
-  Expr new_x_ = em.mkVar("new_x_", bitvector32); // x after executing code (1) or (2)
+  Term new_x = slv.mkConst(bitvector32, "new_x");  // x after executing code (0)
+  Term new_x_ =
+      slv.mkConst(bitvector32, "new_x_");  // x after executing code (1) or (2)
 
   // Encoding code (0)
   // new_x = x == a ? b : a;
-  Expr ite = em.mkExpr(kind::ITE, x_eq_a, b, a);
-  Expr assignment0 = em.mkExpr(kind::EQUAL, new_x, ite);
+  Term ite = slv.mkTerm(ITE, x_eq_a, b, a);
+  Term assignment0 = slv.mkTerm(EQUAL, new_x, ite);
 
   // Assert the encoding of code (0)
   cout << "Asserting " << assignment0 << " to CVC4 " << endl;
-  smt.assertFormula(assignment0);
+  slv.assertFormula(assignment0);
   cout << "Pushing a new context." << endl;
-  smt.push();
+  slv.push();
 
   // Encoding code (1)
   // new_x_ = a xor b xor x
-  Expr a_xor_b_xor_x = em.mkExpr(kind::BITVECTOR_XOR, a, b, x);
-  Expr assignment1 = em.mkExpr(kind::EQUAL, new_x_, a_xor_b_xor_x);
+  Term a_xor_b_xor_x = slv.mkTerm(BITVECTOR_XOR, a, b, x);
+  Term assignment1 = slv.mkTerm(EQUAL, new_x_, a_xor_b_xor_x);
 
   // Assert encoding to CVC4 in current context;
   cout << "Asserting " << assignment1 << " to CVC4 " << endl;
-  smt.assertFormula(assignment1);
-  Expr new_x_eq_new_x_ = em.mkExpr(kind::EQUAL, new_x, new_x_);
+  slv.assertFormula(assignment1);
+  Term new_x_eq_new_x_ = slv.mkTerm(EQUAL, new_x, new_x_);
 
-  cout << " Querying: " << new_x_eq_new_x_ << endl;
-  cout << " Expect entailed. " << endl;
-  cout << " CVC4: " << smt.checkEntailed(new_x_eq_new_x_) << endl;
+  cout << " Check entailment assuming: " << new_x_eq_new_x_ << endl;
+  cout << " Expect ENTAILED. " << endl;
+  cout << " CVC4: " << slv.checkEntailed(new_x_eq_new_x_) << endl;
   cout << " Popping context. " << endl;
-  smt.pop();
+  slv.pop();
 
   // Encoding code (2)
   // new_x_ = a + b - x
-  Expr a_plus_b = em.mkExpr(kind::BITVECTOR_PLUS, a, b);
-  Expr a_plus_b_minus_x = em.mkExpr(kind::BITVECTOR_SUB, a_plus_b, x);
-  Expr assignment2 = em.mkExpr(kind::EQUAL, new_x_, a_plus_b_minus_x);
+  Term a_plus_b = slv.mkTerm(BITVECTOR_PLUS, a, b);
+  Term a_plus_b_minus_x = slv.mkTerm(BITVECTOR_SUB, a_plus_b, x);
+  Term assignment2 = slv.mkTerm(EQUAL, new_x_, a_plus_b_minus_x);
 
   // Assert encoding to CVC4 in current context;
   cout << "Asserting " << assignment2 << " to CVC4 " << endl;
-  smt.assertFormula(assignment2);
+  slv.assertFormula(assignment2);
 
-  cout << " Querying: " << new_x_eq_new_x_ << endl;
+  cout << " Check entailment assuming: " << new_x_eq_new_x_ << endl;
   cout << " Expect ENTAILED. " << endl;
-  cout << " CVC4: " << smt.checkEntailed(new_x_eq_new_x_) << endl;
+  cout << " CVC4: " << slv.checkEntailed(new_x_eq_new_x_) << endl;
 
-  Expr x_neq_x = em.mkExpr(kind::EQUAL, x, x).notExpr();
-  std::vector<Expr> v{new_x_eq_new_x_, x_neq_x};
-  cout << " Querying: " << v << endl;
+  Term x_neq_x = slv.mkTerm(EQUAL, x, x).notTerm();
+  std::vector<Term> v{new_x_eq_new_x_, x_neq_x};
+  cout << " Check entailment assuming: " << v << endl;
   cout << " Expect NOT_ENTAILED. " << endl;
-  cout << " CVC4: " << smt.checkEntailed(v) << endl;
+  cout << " CVC4: " << slv.checkEntailed(v) << endl;
 
-  // Assert that a is odd 
-  Expr extract_op = em.mkConst(BitVectorExtract(0, 0));
-  Expr  lsb_of_a = em.mkExpr(extract_op, a);
-  cout << "Type of " << lsb_of_a << " is " << lsb_of_a.getType() << endl;
-  Expr a_odd = em.mkExpr(kind::EQUAL, lsb_of_a, em.mkConst(BitVector(1u, 1u)));
+  // Assert that a is odd
+  Op extract_op = slv.mkOp(BITVECTOR_EXTRACT, 0, 0);
+  Term lsb_of_a = slv.mkTerm(extract_op, a);
+  cout << "Sort of " << lsb_of_a << " is " << lsb_of_a.getSort() << endl;
+  Term a_odd = slv.mkTerm(EQUAL, lsb_of_a, slv.mkBitVector(1u, 1u));
   cout << "Assert " << a_odd << endl;
   cout << "Check satisfiability." << endl;
-  smt.assertFormula(a_odd);
+  slv.assertFormula(a_odd);
   cout << " Expect sat. " << endl;
-  cout << " CVC4: " << smt.checkSat() << endl;
+  cout << " CVC4: " << slv.checkSat() << endl;
   return 0;
 }
