+++ examples/api/helloworld.cpp
@@ -1,8 +1,8 @@
 /*********************                                                        */
-/*! \file helloworld.cpp
+/*! \file helloworld-new.cpp
  ** \verbatim
  ** Top contributors (to current version):
- **   Tim King, Aina Niemetz
+ **   Aina Niemetz
  ** This file is part of the CVC4 project.
  ** Copyright (c) 2009-2020 by the authors listed in the file AUTHORS
  ** in the top-level source directory) and their institutional affiliations.
@@ -16,14 +16,15 @@
 
 #include <iostream>
 
-#include <cvc4/cvc4.h>
+#include <cvc4/api/cvc4cpp.h>
 
-using namespace CVC4;
-int main() {
-  ExprManager em;
-  Expr helloworld = em.mkVar("Hello World!", em.booleanType());
-  SmtEngine smt(&em);
-  std::cout << helloworld << " is " << smt.checkEntailed(helloworld)
+using namespace CVC4::api;
+
+int main()
+{
+  Solver slv;
+  Term helloworld = slv.mkVar(slv.getBooleanSort(), "Hello World!");
+  std::cout << helloworld << " is " << slv.checkEntailed(helloworld)
             << std::endl;
   return 0;
 }
