+++ examples/api/linear_arith.cpp
@@ -1,8 +1,8 @@
 /*********************                                                        */
-/*! \file linear_arith.cpp
+/*! \file linear_arith-new.cpp
  ** \verbatim
  ** Top contributors (to current version):
- **   Tim King, Aina Niemetz
+ **   Aina Niemetz
  ** This file is part of the CVC4 project.
  ** Copyright (c) 2009-2020 by the authors listed in the file AUTHORS
  ** in the top-level source directory) and their institutional affiliations.
@@ -17,66 +17,66 @@
 
 #include <iostream>
 
-#include <cvc4/cvc4.h>
+#include "cvc4/api/cvc4cpp.h"
 
 using namespace std;
-using namespace CVC4;
+using namespace CVC4::api;
 
-int main() {
-  ExprManager em;
-  SmtEngine smt(&em);
-  smt.setLogic("QF_LIRA"); // Set the logic
+int main()
+{
+  Solver slv;
+  slv.setLogic("QF_LIRA"); // Set the logic
 
   // Prove that if given x (Integer) and y (Real) then
   // the maximum value of y - x is 2/3
 
-  // Types
-  Type real = em.realType();
-  Type integer = em.integerType();
+  // Sorts
+  Sort real = slv.getRealSort();
+  Sort integer = slv.getIntegerSort();
 
   // Variables
-  Expr x = em.mkVar("x", integer);
-  Expr y = em.mkVar("y", real);
+  Term x = slv.mkConst(integer, "x");
+  Term y = slv.mkConst(real, "y");
 
   // Constants
-  Expr three = em.mkConst(Rational(3));
-  Expr neg2 = em.mkConst(Rational(-2));
-  Expr two_thirds = em.mkConst(Rational(2,3));
+  Term three = slv.mkReal(3);
+  Term neg2 = slv.mkReal(-2);
+  Term two_thirds = slv.mkReal(2, 3);
 
   // Terms
-  Expr three_y = em.mkExpr(kind::MULT, three, y);
-  Expr diff = em.mkExpr(kind::MINUS, y, x);
+  Term three_y = slv.mkTerm(MULT, three, y);
+  Term diff = slv.mkTerm(MINUS, y, x);
 
   // Formulas
-  Expr x_geq_3y = em.mkExpr(kind::GEQ, x, three_y);
-  Expr x_leq_y = em.mkExpr(kind::LEQ, x, y);
-  Expr neg2_lt_x = em.mkExpr(kind::LT, neg2, x);
+  Term x_geq_3y = slv.mkTerm(GEQ, x, three_y);
+  Term x_leq_y = slv.mkTerm(LEQ, x, y);
+  Term neg2_lt_x = slv.mkTerm(LT, neg2, x);
 
-  Expr assumptions =
-    em.mkExpr(kind::AND, x_geq_3y, x_leq_y, neg2_lt_x);
+  Term assertions =
+    slv.mkTerm(AND, x_geq_3y, x_leq_y, neg2_lt_x);
 
-  cout << "Given the assumptions " << assumptions << endl;
-  smt.assertFormula(assumptions);
+  cout << "Given the assertions " << assertions << endl;
+  slv.assertFormula(assertions);
 
 
-  smt.push();
-  Expr diff_leq_two_thirds = em.mkExpr(kind::LEQ, diff, two_thirds);
+  slv.push();
+  Term diff_leq_two_thirds = slv.mkTerm(LEQ, diff, two_thirds);
   cout << "Prove that " << diff_leq_two_thirds << " with CVC4." << endl;
   cout << "CVC4 should report ENTAILED." << endl;
-  cout << "Result from CVC4 is: " << smt.checkEntailed(diff_leq_two_thirds)
+  cout << "Result from CVC4 is: " << slv.checkEntailed(diff_leq_two_thirds)
        << endl;
-  smt.pop();
+  slv.pop();
 
   cout << endl;
 
-  smt.push();
-  Expr diff_is_two_thirds = em.mkExpr(kind::EQUAL, diff, two_thirds);
-  smt.assertFormula(diff_is_two_thirds);
-  cout << "Show that the asserts are consistent with " << endl;
+  slv.push();
+  Term diff_is_two_thirds = slv.mkTerm(EQUAL, diff, two_thirds);
+  slv.assertFormula(diff_is_two_thirds);
+  cout << "Show that the assertions are consistent with " << endl;
   cout << diff_is_two_thirds << " with CVC4." << endl;
   cout << "CVC4 should report SAT." << endl;
-  cout << "Result from CVC4 is: " << smt.checkSat(em.mkConst(true)) << endl;
-  smt.pop();
+  cout << "Result from CVC4 is: " << slv.checkSat() << endl;
+  slv.pop();
 
   cout << "Thus the maximum value of (y - x) is 2/3."<< endl;
